Fix the defects found while importing the in-service codes
The previous commit reproduced the in-service hex exactly, bugs included.
This one makes each code do what it says.

Netplay: Bobble More, Random Batting Power, Random Star Pitch and Random Star
Swing rolled on the CPU timebase, which differs per client. They now use the
game's seeded RandomInt_Game, became C bodies, and no longer trash registers.
Bobble More is the advertised 50% (the hex worked out to 25%) and no longer
clobbers cr0 ahead of the minigame branch. Random Star Swing never rolls 0
("no captain swing").

Input Prediction and Batter Lag Reduction share one batting implementation
(Match/BattingPrediction.h); both are kept. Input Prediction's pitch
prediction read port 1's controller for every pitcher and now reads the
fielding team's port, skipping CPU pitchers.

Also fixed: Hold Up for Eephus fired on Z/R/L; Ban Characters on CSS wrote
out of bounds with no cursor on the grid; Always 0 outs and Reset Stars in
13th Inning only looked at low bytes; Hazardless w/ Star Pads patched every
stadium; Remove Slice - Frame 2 Stars carried a stray entry. Balance Patch
and the April Fools patch now match their change logs (Blue Shy Guy Speed,
Dry Bones keeps Wall Splat, Red Koopa / Petey stat bars, no r31 clobber), and
the Toad hitbox outlier is 350 like the other Toads.

Names follow behaviour: Ranked/Ban Batter Pausing.c held the "restrict"
logic and is now Restrict Batter Pausing.c; the outright ban is
Match/Ban Batter Pausing.c.

+++ Gecko Codes/Match/Reset Stars in 13th Inning.c
@@ -18,7 +18,7 @@ void ResetStarsIn13thInning(void)
     READ_GAME_REG(s32, finishedInning, 3);
 
     g_Scores.Inning = finishedInning + 1;
-    if ((u8)g_Scores.Inning != RESET_INNING)
+    if (g_Scores.Inning != RESET_INNING)
         return;
 
     g_GameLogic.TeamStars[0] = Static_Stats_Tables.startingChemStars[0];
